+++ src/mdc-ripple/directive.js
@@ -10,33 +10,21 @@ import {MDCRippleMixin} from './mixin';
  * @module mdc.ripple
  * @description Apply a ripple to the given element when clicked.
  *
- * @param {bool} [mdcRippleIsUnbounded] - override the unbounded property
+ * @param {bool} [dataMdcRippleIsUnbounded] - override the unbounded property
  */
-export class MDCRippleController extends MDCRippleMixin(BindInjections) { /* eslint new-cap: 0 */
+export class MDCRippleController extends MDCRippleMixin(BindInjections) {
   static get name() {
     return 'mdcRipple';
   }
 
-  static get bindings() {
-    return {
-      mdcRippleIsUnbounded: '<?',
-    };
-  }
-
   static get $inject() {
-    return arrayUnion(super.$inject, ['$scope']);
+    return arrayUnion(super.$inject, ['$element']);
   }
 
   constructor(...args) {
     super(...args);
 
     this.$element.addClass('mdc-ripple-surface');
-
-    this.$scope.$watch('mdcRippleIsUnbounded', () => {
-      if (this.$scope.mdcRippleIsUnbounded !== undefined) {
-        this.ripple.unbounded = this.$scope.mdcRippleIsUnbounded;
-      }
-    });
   }
 }
 
@@ -51,5 +39,5 @@ angular
   .module('mdc.ripple', [])
   .directive(MDCRippleController.name, () => ({
     controller: MDCRippleController,
-    scope: MDCRippleController.bindings,
+    priority: Number.MIN_SAFE_INTEGER,
   }));