+++ src/mdc-ripple/mixin.js
@@ -6,18 +6,31 @@ import {MDCRipple} from '@material/ripple';
 /**
  * Applies a ripple to this.$element and exposes as this.ripple.
  *
- * @mixin HasMDCRipple
+ * @mixin MDCRippleMixin
  * @param Base - should extend BindInjections or a subclass thereof
  */
 export const MDCRippleMixin = (Base) => class extends Base {
   static get $inject() {
-    return arrayUnion(super.$inject, ['$element']);
+    return arrayUnion(['$element'], super.$inject);
   }
 
   constructor(...args) {
     super(...args);
 
-    this.ripple = MDCRipple.attachTo(this.$element[0]);
+    this.ripple = {};
+    this.$element.ready(() => {
+      if (this.ripple.doDestroy) {
+        return;
+      }
+      // if unbounded is assigned before ready, we pass it in
+      this.ripple = MDCRipple.attachTo(this.$element[0], {isUnbounded: this.ripple.unbounded});
+    });
+  }
+
+  $postLink() {
+    if (super.$postLink) {
+      super.$postLink();
+    }
   }
 
   $onDestroy() {
@@ -25,8 +38,10 @@ export const MDCRippleMixin = (Base) => class extends Base {
       super.$onDestroy();
     }
 
-    if (this.ripple) {
+    if (this.ripple.destroy) {
       this.ripple.destroy();
+    } else {
+      this.ripple.doDestroy = true;
     }
   }
 };